fix: gen-submodule-env when submodule uninitialized

+++ gw_util.sh
@@ -123,14 +123,22 @@ get_sudt_code_hash_from_lumos_file() {
     echo "$(cat $lumosconfigfile)" | grep -Pzo 'SUDT[\s\S]*CODE_HASH": "\K[^"]*'
 }
 
-# 
+ 
 generateSubmodulesEnvFile(){
     File="docker/.manual.build.list.env"
-    rm $File 
+    if [[ -f $File ]]; then
+        rm $File 
+    fi
     echo "####[mode]" >> $File
     echo MANUAL_BUILD_GODWOKEN=false >> $File
     echo MANUAL_BUILD_WEB3=false >> $File
     echo '' >> $File
+
+    # if submodule folder is not initialized and updated
+    if [[ -z "$(ls -A godwoken)" || -z "$(ls -A godwoken-examples)" || -z "$(ls -A godwoken-polyjuice)" || -z "$(ls -A godwoken-web3)" ]]; then
+       echo "one or more of submodule folders is Empty, do init and update first."
+       git submodule update --init --recursive
+    fi
 
     local -a arr=("godwoken" "godwoken-web3" "godwoken-polyjuice" "godwoken-examples")
     for i in "${arr[@]}"
@@ -144,7 +152,7 @@ generateSubmodulesEnvFile(){
        # get last commit
        commit=$(git config --file .gitmodules --get-regexp "submodule.${i}.path" | 
         awk '{print $2}' | xargs -i git -C {} log --pretty=format:'%h' -n 1 )
-    
+
        # renameing godwoken-examples => godwoken_examples, 
        # cater for env variable naming rule.
        url_name=$(echo "${i^^}_URL" | tr - _ )
